clean up decompressor tests
