+++ test/helpers/codec/index.ts
@@ -1,1 +1,2 @@
 export * from './revert-flags'
+export * from './encoding'
